+++ noodles-sam/src/record/data/field/value/ty.rs
@@ -11,8 +11,18 @@ use std::{
 pub enum Type {
     /// Character (`A`).
     Char,
-    /// Integer (`i`).
+    /// 8-bit integer (`c`).
+    Int8,
+    /// 8-bit unsigned integer (`C`).
+    UInt8,
+    /// 16-bit integer (`s`).
+    Int16,
+    /// 16-bit unsigned integer (`S`).
+    UInt16,
+    /// 32-bit integer (`i`).
     Int32,
+    /// 32-bit unsigned integer (`I`).
+    UInt32,
     /// Single-precision floating-point (`f`).
     Float,
     /// String (`Z`).
@@ -39,7 +49,7 @@ impl fmt::Display for ParseError {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(
             f,
-            "invalid data field type: expected {{A, i, f, Z, H, B}}, got {}",
+            "invalid data field type: expected {{A, c, C, s, S, i, I, f, Z, H, B}}, got {}",
             self.0
         )
     }
@@ -51,7 +61,12 @@ impl FromStr for Type {
     fn from_str(s: &str) -> Result<Self, Self::Err> {
         match s {
             "A" => Ok(Self::Char),
+            "c" => Ok(Self::Int8),
+            "C" => Ok(Self::UInt8),
+            "s" => Ok(Self::Int16),
+            "S" => Ok(Self::UInt16),
             "i" => Ok(Self::Int32),
+            "I" => Ok(Self::UInt32),
             "f" => Ok(Self::Float),
             "Z" => Ok(Self::String),
             "H" => Ok(Self::Hex),
@@ -65,7 +80,12 @@ impl From<Type> for char {
     fn from(ty: Type) -> Self {
         match ty {
             Type::Char => 'A',
+            Type::Int8 => 'c',
+            Type::UInt8 => 'C',
+            Type::Int16 => 's',
+            Type::UInt16 => 'S',
             Type::Int32 => 'i',
+            Type::UInt32 => 'I',
             Type::Float => 'f',
             Type::String => 'Z',
             Type::Hex => 'H',
@@ -81,7 +101,12 @@ mod tests {
     #[test]
     fn test_fmt() {
         assert_eq!(Type::Char.to_string(), "A");
+        assert_eq!(Type::Int8.to_string(), "c");
+        assert_eq!(Type::UInt8.to_string(), "C");
+        assert_eq!(Type::Int16.to_string(), "s");
+        assert_eq!(Type::UInt16.to_string(), "S");
         assert_eq!(Type::Int32.to_string(), "i");
+        assert_eq!(Type::UInt32.to_string(), "I");
         assert_eq!(Type::Float.to_string(), "f");
         assert_eq!(Type::String.to_string(), "Z");
         assert_eq!(Type::Hex.to_string(), "H");
@@ -91,7 +116,12 @@ mod tests {
     #[test]
     fn test_from_str() {
         assert_eq!("A".parse(), Ok(Type::Char));
+        assert_eq!("c".parse(), Ok(Type::Int8));
+        assert_eq!("C".parse(), Ok(Type::UInt8));
+        assert_eq!("s".parse(), Ok(Type::Int16));
+        assert_eq!("S".parse(), Ok(Type::UInt16));
         assert_eq!("i".parse(), Ok(Type::Int32));
+        assert_eq!("I".parse(), Ok(Type::UInt32));
         assert_eq!("f".parse(), Ok(Type::Float));
         assert_eq!("Z".parse(), Ok(Type::String));
         assert_eq!("H".parse(), Ok(Type::Hex));
@@ -108,7 +138,12 @@ mod tests {
     #[test]
     fn test_from_type_for_char() {
         assert_eq!(char::from(Type::Char), 'A');
+        assert_eq!(char::from(Type::Int8), 'c');
+        assert_eq!(char::from(Type::UInt8), 'C');
+        assert_eq!(char::from(Type::Int16), 's');
+        assert_eq!(char::from(Type::UInt16), 'S');
         assert_eq!(char::from(Type::Int32), 'i');
+        assert_eq!(char::from(Type::UInt32), 'I');
         assert_eq!(char::from(Type::Float), 'f');
         assert_eq!(char::from(Type::String), 'Z');
         assert_eq!(char::from(Type::Hex), 'H');