+++ noodles-sam/CHANGELOG.md
@@ -20,6 +20,14 @@
 
   * sam/reader: Implement `AlignmentReader`.
 
+  * sam/record/data/field/value: Add integer types.
+
+    This means that SAM data field integer values can be typed, i.e., it adds
+    the `Int8` (`c`), `UInt8` (`C`), `Int16` (`s`), `UInt16` (`S`), and
+    `UInt32` (`I`) types. Though these aren't actually used in the SAM format,
+    they can transparently be used in SAM and shared among the different
+    alignment formats.
+
   * sam/record/cigar: Add `clear` method.
 
   * sam/record/cigar: Implement `AsMut<Vec<Op>>`.